requirements.txt
Django==1.9.9
environs==4.2.0
django-tinymce==2.6.0
pillow==6.2.2
